Moved the code from the current <Popover> into <LegacyPopover> which is rendered if an array of elements is passed to <Popover>. Added deprecation warnings for deprecated props. Refactored the <Popover> to use <Popper> and added various options for rendering the popover.
