+++ lib/InfoPopover/InfoPopover.js
@@ -8,7 +8,7 @@ import classNames from 'classnames';
 import { FormattedMessage } from 'react-intl';
 import noop from 'lodash/noop';
 
-import Popover, { PopoverConsumer } from '../Popover';
+import Popover, { PopoverConsumer } from '../Popover/LegacyPopover/LagacyPopover';
 import Button from '../Button';
 import IconButton from '../IconButton';
