+++ lib/Popover/stories/Popover.stories.js
@@ -3,7 +3,9 @@ import { storiesOf } from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
 import readme from '../readme.md';
 import BasicUsage from './BasicUsage';
+import Refactor from './refactor';
 
 storiesOf('Popover', module)
   .addDecorator(withReadme(readme))
-  .add('Basic Usage', () => <BasicUsage />);
+  .add('Basic Usage', () => <BasicUsage />)
+  .add('Refactor', () => <Refactor />);
